+++ test/test_edicao.py
@@ -1,7 +1,6 @@
 from unittest.mock import patch
 
 import pytest
-from rich.console import Console
 
 from .db_test import criar_banco_teste_com_dados
 from .db_test import remove_banco_teste
@@ -11,7 +10,6 @@ from trackJobs.banco_de_dados import BancoDeDados
 from trackJobs.edicao import edicao
 from trackJobs.edicao import MenuEdicao
 from trackJobs.menu import Menu
-from trackJobs.validador import VALIDADORES
 
 
 def get_vaga_id(db_path, id):
@@ -76,30 +74,4 @@ def test_edicao(mock_menu, campo_alterado, novo_valor, msg_esperada, request):
     candidatura_alterada = get_vaga_id("track_jobs_test.db", id_da_vaga_escolhida)
     assert candidatura_alterada[f"{campo_alterado}"] == novo_valor
 
-    remove_banco_teste()
-
-
-@pytest.mark.parametrize(
-    "campo_selecionado, novo_dado, esperado_valido",
-    [
-        ("nome", "", False),  # Nome vazio
-        ("nome", "Desenvolvedor Backend", True),  # Nome válido
-        ("link", "invalid-url", False),  # URL inválida
-        ("link", "https://jobs.example.com/valid", True),  # URL válida
-        ("status", "inexistente", False),  # Status inválido
-        ("status", "aceito", True),  # Status válido
-        ("data de aplicaçao", "2021-02-30", False),  # Data inválida
-        ("data de aplicaçao", "2023-10-10", True),  # Data válida
-        ("descriçao", "", True),  # Descrição sempre válida
-    ],
-)
-def test_validacao_dados_edicao(campo_selecionado, novo_dado, esperado_valido):
-    criar_banco_teste_com_dados()
-    with patch("questionary.text") as mock_text:
-        mock_text.return_value.ask.return_value = novo_dado
-        console = Console()
-        valido = VALIDADORES[campo_selecionado](
-            "track_jobs_test.db", novo_dado, console
-        )
-        assert valido == esperado_valido
     remove_banco_teste()
